requirements.txt
-i https://developer.download.nvidia.com/compute/redist
nvidia-dali-cuda110
lmdb
pytest
munch
pyyaml